requirements.txt
Flask
requests
openai
discord.py
pymongo
python-dotenv
discord
matplotlib
numpy
urllib3<2.0
plotly
kaleido
beautifulsoup4
joblib
pandas
ta
scikit-learn
yfinance